+++ NEWS.md
@@ -1,3 +1,8 @@
+vivo 0.1.2
+----------------------------------------------------------------
+* fix in `plot()` function, change in `DALEX` package
+* small fixes to `local_variable_importance()`
+
 vivo 0.1.1
 ----------------------------------------------------------------
 * fix in `calculate_weight()`, change in `ingredients` package